some correction

+++ task1_solution/src/file.ts
@@ -4,7 +4,6 @@ import * as fs from 'fs-extra';
 import * as https from 'https';
 import { WriteStream } from 'fs';
 
-
 export const FileUtils = {
   createTempFilePath,
   createTempFile,
@@ -16,71 +15,72 @@ export const FileUtils = {
   ensureFolder,
   writeFile,
   readFile,
+  tempFileFinder,
 };
 
+function tempFileFinder(){
+  return os.tmpdir()
+}
 
-
-
- function createTempFilePath(fileName: string, folderPath?: string): string {
+function createTempFilePath(fileName: string, folderPath?: string): string {
   const components = folderPath ? [...folderPath.split('/'), fileName] : [fileName];
   const tempFilePath = path.join(os.tmpdir(), ...components);
-
   return tempFilePath;
 }
 
- function createTempFile(tempFilePath: string): WriteStream {
+function createTempFile(tempFilePath: string): WriteStream {
   if (!tempFilePath.includes(os.tmpdir())) {
     throw new Error('Path should be inside tmp directory');
   }
   return fs.createWriteStream(tempFilePath);
 }
 
- async function downloadTempFile(url: string, destinationPath: string): Promise<void> {
+async function downloadTempFile(url: string, destinationPath: string): Promise<void> {
   const file = createTempFile(destinationPath);
   https.get(url, (result) => {
     result.pipe(file);
   });
-
   await new Promise((resolve, reject) => {
     file.on('finish', resolve);
     file.on('error', reject);
   });
 }
 
- function ensureFolder(tempFolderPath: string): void {
+function ensureFolder(tempFolderPath: string): void {
   if (!fs.existsSync(tempFolderPath)) {
     fs.mkdirSync(tempFolderPath, { recursive: true });
   }
 }
 
- function getFolderPath(filePath: string): string {
+function getFolderPath(filePath: string): string {
   return path.dirname(filePath);
 }
 
- function getFileName(filePath: string): string {
+function getFileName(filePath: string): string {
   return path.basename(filePath);
 }
 
-
- function removeTempFile(tempFilePath: string) {
+function removeTempFile(tempFilePath: string) {
   fs.unlinkSync(tempFilePath);
 }
 
- function writeFile(fileNamePath: string, data: any) {
+function writeFile(fileNamePath: string, data: any) {
   const dirname = path.dirname(fileNamePath);
   if (dirname) {
     ensureFolder(dirname);
   }
-
   fs.writeFile(fileNamePath, JSON.stringify(data), function (err) {
     if (err) throw err;
   });
 }
 
- function readFile(fileNamePath: string): Buffer {
+function readFile(fileNamePath: string): Buffer {
   return fs.readFileSync(fileNamePath);
 }
- function removeDir(dirPath: string): Promise<void> {
-  return fs.emptyDir(dirPath);
+
+function removeDir(dirPath: string): Promise<void> {
+  if (dirPath.includes(os.tmpdir())) {
+    return fs.emptyDir(dirPath);
+  }
 }
